chore: cargo clippy --fix mechanical sweep
- manual_div_ceil: (n + d - 1) / d -> n.div_ceil(d) (4 sites).
- redundant_locals / unnecessary_cast cleanups.
- items_after_test_module: move default_seed before mod tests in state.rs.
- derive(Default) collapses where the impl was trivial (moe_experts.rs Activation).
- Misc tiny mechanical rewrites flagged by clippy 1.96 across 19 files.

+++ crates/backend-hip/tests/peer_access_pairs.rs
@@ -19,7 +19,7 @@ fn maybe_skip() -> Option<i32> {
 #[test]
 fn pair_matrix_each_two_device_subset() {
     let Some(n) = maybe_skip() else { return };
-    let n = n as i32;
+    let n = n;
     if n < 2 {
         eprintln!("[skip] need >= 2 devices (have {n})");
         return;
